Layout configuration improvements, and configuration code.
The configuration code was under an extension, but it was used by all kinds of parts of the system that did not care about that extension.  So the code has been moved out into the "base" package.  It is not necessary to be in base, but it does not belong in aid, and it is so ubiquitous that it is good enough for now.