Migrate validation code.
The old v3 validation code had some nice ideas, but it was used too sparingly in the old codebase.  With the new ideas around the core, this may be used more often.  It will still run into the issues that mypy doesn't know about them, so there may be some duplication of effort.