Fixed major problem with the extension loader, where it wasn't waiting on implementations of extensions to finish loading if the dependency was an API.
This was fixed by adding in a new event that all extensions must send to indicate that it finished loading.  This avoids really tricky and troublesome on-startup code that only affects an extension at startup.

There's still a bug being hunted down to get the hotkey binding to correctly set the native hotkey binding state.